Refactored ExportReader
